Introduce AbstractDnfStateBasedDomain as subclass of AbstractStateBasedDomain
