+++ trunk/source/Library-Sifa/src/de/uni_freiburg/informatik/ultimate/lib/sifa/domain/OctagonDomain.java
@@ -47,7 +47,7 @@ import de.uni_freiburg.informatik.ultimate.logic.TermTransformer;
  * @author Frank Schüssele (schuessf@informatik.uni-freiburg.de)
  *
  */
-public class OctagonDomain extends AbstractStateBasedDomain<OctagonState> {
+public class OctagonDomain extends AbstractDnfStateBasedDomain<OctagonState> {
 	private final TermTransformer mTermTransformer = new RewriteEqualityTransformer(mTools.getScript());
 
 	public OctagonDomain(final ILogger logger, final SymbolicTools tools, final int maxDisjuncts,
